Support #11237 Support #11221: Anexo 14 - Nuevo Validador C.2.B ex #11042 Anexo 14 - Nuevo Validador C.2.B ex #11042

+++ application/modules/sgr/helpers/tools_helper.php
@@ -876,13 +876,13 @@ function translate_month_spanish($code) {
 
 function fix_ten_cents($value, $value2) {
 
-    $percent = $value * 0.01;
-    $percent_negative = $value * -0.01;
+    $percent = (int) ($value * 0.01);
+    $percent_negative = (int)($value * -0.01);
 
     $range = range($percent_negative, $percent);
     $diff = $value - $value2;
     $diff = (int) $diff;
-
+    
     if (!in_array($diff, $range))
         return true;
 }
